refactor(tests): deduplicate _enable_log_propagation fixture + extract prep_harness
Move _enable_log_propagation to shared conftest (was duplicated in
Phase 1 and Phase 2 test files). Extract repeated test setup in
TestContextMapIntegrity into a prep_harness fixture and _run helper.
