+++ tests/unit/unification/test_phase2_batch_prep_integrity.py
@@ -23,20 +23,30 @@ from agent_actions.record.state import RecordState
 _PREPARATOR_LOGGER = "agent_actions.llm.batch.processing.preparator"
 
 
-@pytest.fixture(autouse=True)
-def _enable_log_propagation():
-    """Ensure agent_actions loggers propagate to root so caplog captures them."""
-    aa_logger = logging.getLogger("agent_actions")
-    orig = aa_logger.propagate
-    aa_logger.propagate = True
-    yield
-    aa_logger.propagate = orig
-
-
 class TestContextMapIntegrity:
     """U-2.B: context_map must not show INCLUDED for records where prepare() failed."""
 
-    def test_prepare_failure_not_marked_included(self, batch_preparator):
+    @pytest.fixture()
+    def prep_harness(self, batch_preparator):
+        """Shared setup for _process_single_item tests."""
+        return {
+            "preparator": batch_preparator,
+            "context_map": {},
+            "stats": BatchTaskPreparationStats(),
+            "prep_context": MagicMock(spec=PreparationContext),
+        }
+
+    def _run(self, prep_harness, row, task_preparer):
+        """Run _process_single_item with the shared harness."""
+        return prep_harness["preparator"]._process_single_item(
+            row,
+            prep_harness["prep_context"],
+            task_preparer,
+            prep_harness["context_map"],
+            prep_harness["stats"],
+        )
+
+    def test_prepare_failure_not_marked_included(self, prep_harness):
         """If prepare() throws, context_map must NOT have INCLUDED for that row.
 
         Bug: INCLUDED is set before prepare() runs, so when prepare() throws,
@@ -44,35 +54,24 @@ class TestContextMapIntegrity:
         context_map entry is left with INCLUDED — a lie.
         """
         row = {"target_id": "t-001", "content": {"text": "will fail"}}
-        context_map: dict = {}
-        stats = BatchTaskPreparationStats()
-        prep_context = MagicMock(spec=PreparationContext)
-
         task_preparer = MagicMock()
         task_preparer.prepare.side_effect = ValueError("schema validation failed")
 
         with pytest.raises(ValueError, match="schema validation failed"):
-            batch_preparator._process_single_item(
-                row, prep_context, task_preparer, context_map, stats
-            )
+            self._run(prep_harness, row, task_preparer)
 
-        # Row must be in context_map (for _mark_prep_failed to find it later)
+        context_map = prep_harness["context_map"]
         assert "t-001" in context_map
 
-        # But it must NOT be marked INCLUDED — prepare() never succeeded
         status = BatchContextMetadata.get_filter_status(context_map["t-001"])
         assert status != FilterStatus.INCLUDED, (
             f"context_map shows INCLUDED for row where prepare() failed — "
             f"INCLUDED was set before prepare() ran (got status={status})"
         )
 
-    def test_successful_prepare_marked_included(self, batch_preparator):
+    def test_successful_prepare_marked_included(self, prep_harness):
         """When prepare() succeeds and guard passes, context_map shows INCLUDED."""
         row = {"target_id": "t-002", "content": {"text": "will succeed"}}
-        context_map: dict = {}
-        stats = BatchTaskPreparationStats()
-        prep_context = MagicMock(spec=PreparationContext)
-
         task_preparer = MagicMock()
         prepared = MagicMock()
         prepared.guard_status = GuardStatus.PASSED
@@ -81,35 +80,27 @@ class TestContextMapIntegrity:
         prepared.formatted_prompt = "Process this"
         task_preparer.prepare.return_value = prepared
 
-        result = batch_preparator._process_single_item(
-            row, prep_context, task_preparer, context_map, stats
-        )
+        result = self._run(prep_harness, row, task_preparer)
 
         assert result is not None
-        status = BatchContextMetadata.get_filter_status(context_map["t-002"])
+        status = BatchContextMetadata.get_filter_status(prep_harness["context_map"]["t-002"])
         assert status == FilterStatus.INCLUDED, (
             f"context_map must show INCLUDED after successful prepare (got status={status})"
         )
 
-    def test_guard_skipped_not_marked_included(self, batch_preparator):
+    def test_guard_skipped_not_marked_included(self, prep_harness):
         """When guard skips a record, context_map shows SKIPPED, not INCLUDED."""
         row = {"target_id": "t-003", "content": {"text": "guard skip"}}
-        context_map: dict = {}
-        stats = BatchTaskPreparationStats()
-        prep_context = MagicMock(spec=PreparationContext)
-
         task_preparer = MagicMock()
         prepared = MagicMock()
         prepared.guard_status = GuardStatus.SKIPPED
         prepared.passthrough_fields = {}
         task_preparer.prepare.return_value = prepared
 
-        result = batch_preparator._process_single_item(
-            row, prep_context, task_preparer, context_map, stats
-        )
+        result = self._run(prep_harness, row, task_preparer)
 
         assert result is None
-        status = BatchContextMetadata.get_filter_status(context_map["t-003"])
+        status = BatchContextMetadata.get_filter_status(prep_harness["context_map"]["t-003"])
         assert status == FilterStatus.SKIPPED
